Add comprehensive integration and unit tests for services and endpoints
